moved all constants to utils and updated them
