+++ src/main/java/com/online_store/utils/constant/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.online_store.constants;
+package com.online_store.utils.constant;
 
 public class ErrorMessage {
     public static final String EMAIL_NOT_FOUND = "Email not found";
@@ -16,6 +16,14 @@ public class ErrorMessage {
     public static final String CATEGORY_NOT_FOUND = "Category not found";
     public static final String INVALID_CATEGORY_NAME = "Category with such name doesn't exist";
     public static final String PRODUCT_NOT_FOUND = "Category not found";
+    public static final String USER_NOT_ACTIVE = "User's email address has not been confirmed.";
+    public static final String PASSWORDS_CANT_BE_EMPTY = "Passwords can't be empty";
+    public static final String INCORRECT_OLD_PASSWORD = "Incorrect old password";
+    // for JWT
+    public static final String INVALID_REFRESH_TOKEN = "Only refresh token is allowed";
+    public static final String MISS_BEARER_PREFIX = "Missing 'Bearer' type in 'Authorization' header. Expected 'Authorization: Bearer <JWT>'";
+    public static final String TOKEN_EXPIRED = "JWT token is expired";
+    public static final String TOKEN_INVALID = "JWT token is invalid";
 
     private ErrorMessage() {
     }
