+++ src/main/java/com/online_store/utils/constant/Path.java
@@ -1,4 +1,4 @@
-package com.online_store.constants;
+package com.online_store.utils.constant;
 
 public class Path {
     public static final String API_VERSION = "/v1";
@@ -7,6 +7,7 @@ public class Path {
     public static final String PRODUCT = API_PATHNAME + "/product";
     public static final String ORDER = API_PATHNAME + "/order";
     public static final String CATEGORY = API_PATHNAME + "/category";
+    public static final String GENERATOR = API_PATHNAME + "/generate";
     public static final String FRONT_END_LOCALHOST = "http://localhost:3000";
     private Path() {
     }
